Move/refactor ApplyTensorYlmFilter.
This is to make dependencies nicer:
now NumericalAlgorithms/SphericalHarmonics does not
depend on GeneralizedHarmonic or CurvedScalarWave.

No nontrivial changes in actual code; code was just moved from
NumericalAlgorithms/SphericalHarmonics into GeneralizedHarmonic
and CurvedScalarWave, keeping generic code in SphericalHarmonics.

Similarly for tests.
