+++ src/Evolution/Systems/GeneralizedHarmonic/CMakeLists.txt
@@ -8,6 +8,7 @@ add_spectre_library(${LIBRARY})
 spectre_target_sources(
   ${LIBRARY}
   PRIVATE
+  ApplyTensorYlmFilter.cpp
   Characteristics.cpp
   Constraints.cpp
   Equations.cpp
@@ -18,6 +19,7 @@ spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
+  ApplyTensorYlmFilter.hpp
   Characteristics.hpp
   Constraints.hpp
   DuDtTempTags.hpp
