+++ src/NumericalAlgorithms/SphericalHarmonics/CMakeLists.txt
@@ -9,7 +9,6 @@ spectre_target_sources(
   ${LIBRARY}
   PRIVATE
   AngularOrdering.cpp
-  ApplyTensorYlmFilter.cpp
   ChangeCenterOfStrahlkorper.cpp
   RealSphericalHarmonics.cpp
   SpherepackIterator.cpp
@@ -33,6 +32,7 @@ spectre_target_headers(
   HEADERS
   AngularOrdering.hpp
   ApplyTensorYlmFilter.hpp
+  ApplyTensorYlmFilter.tpp
   ChangeCenterOfStrahlkorper.hpp
   RealSphericalHarmonics.hpp
   SpherepackIterator.hpp
