+++ tests/Unit/Evolution/Systems/CurvedScalarWave/CMakeLists.txt
@@ -12,6 +12,7 @@ set(LIBRARY_SOURCES
   BoundaryConditions/Test_ConstraintPreservingSphericalRadiation.cpp
   BoundaryConditions/Test_DemandOutgoingCharSpeeds.cpp
   BoundaryConditions/Test_Worldtube.cpp
+  Test_ApplyTensorYlmFilter.cpp
   Test_BackgroundSpacetime.cpp
   Test_CalculateGrVars.cpp
   Test_Constraints.cpp
