+++ tests/Unit/NumericalAlgorithms/SphericalHarmonics/CMakeLists.txt
@@ -5,7 +5,6 @@ set(LIBRARY "Test_SphericalHarmonics")
 
 set(LIBRARY_SOURCES
   Test_AngularOrdering.cpp
-  Test_ApplyTensorYlmFilter.cpp
   Test_ChangeCenterOfStrahlkorper.cpp
   Test_RealSphericalHarmonics.cpp
   Test_Spherepack.cpp
